2011-03-26, First release of http://enlargeyourpassword.com (no CSS style)
I finally started working on this idea today. I first developped a static
prototype in HTML to get a feel for what it would look like to use the tool.
I then added the generation of passwords using JavaScript.

I initially intended to include only three password sizes (L,XL,XXL) using
three flavors of SHA, but I had to add a shorter size (M) to get a password
of only 32 characters (that was the maximum allowed) for the first service that
I signed up for.

It may be useful to keep a list of which service supports which size, for
information.