Remove jslint comment
The names of parameters have proven to be unstable with the evolution
of jslint. The intent of the use of scope() is to write code in a more
stable way, that will not need to be rewritten for reuse in the future.
In the same vein, the jlint instruction is too anchored in the present;
jslint options should be selected in the tool itself instead.