Move utility functions to a separate script, utils.js
The script utils.js defines several small modules, which are grouped
in a single file to limit the number of HTTP requests.

This is the exception in this project, where I have kept scripts
separate to preserve their traceability, and voluntarily avoided
obfuscation to let users check that nothing dubious is happening:
no third party tracking, no submission to a server.