Use scope() from scope-or-not project to declare modules
The intent of this change is to make the declaration of modules
more consistent and stable through time.

Previously, most modules declared all their code in the global scope,
and one used define() from Asynchronous Module Definition format for
which a stub had to be provided specifically for this module.