Add style for width of inputs
I have been satisfied with the default width of inputs until using
the page in Firefox on Ubuntu 12.04, where inputs are too large
by default and the text of the form had too many line breaks.

I added styles for the width of inputs to get a more consistent
experience across operating systems and browsers.