Trim each collected input value, without updating the input field itself
The motivation for this change is to avoid discrepancies in generated
passwords when nearly invisible space characters are included at the
start or end of a field value. These extraneous space characters are
easily introduced, unbeknownst to the user, when selecting a value to
copy/paste from a separate file or text field.

Following this change, space characters at the start and end of each
field value are silently discarded for the computation of the generated
passwords, but left intact in the value of the input field, which allows
to type a sentence or a string of words including  space characters.