Ignore source Numbers spreadsheet which includes Numbers versioning
Numbers creates versions during editing, which can be reverted to using

  File > Revert To > Browse All Versions...

This is redundant with git versioning.

I tried to disable that behavior by setting the option

  [x] Ask to keep changes when closing documents

in System Preferences but it had no effect.

In order to add a file with only the latest version for git versioning,
I use 'Save As' to save the file with a different name, removing the
.versions part from the file extension.

Note that the 'Save As' option is not visible by default in the File menu.
In order to make it appear, you need to press the Option key.